# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Talib MR - University Student Platform for Mauritania
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout (invalidate refresh token)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/refresh:
    post:
      operationId: refreshTokens
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required: [refreshToken]
      responses:
        "200":
          description: New tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "200":
          description: Reset email queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── UNIVERSITIES & ACADEMIC STRUCTURE ───────────────────────────────────────
...
  /community/posts:
    get:
      operationId: listPosts
      tags: [community]
      summary: List community posts
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Post"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                required: [success, data]

    post:
      operationId: createPost
      tags: [community]
      summary: Create a community post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: "#/components/schemas/Post"
                required: [success, data]
  /community/posts/{postId}/react:
    post:
      operationId: reactToPost
      tags: [community]
      summary: React to a post
      security:
        - bearerAuth: []
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reaction:
                  type: string
                  enum: [like, love, insightful, helpful]
              required: [reaction]
      responses:
        "200":
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── EVENTS ──────────────────────────────────────────────────────────────────
...
  /clubs:
    get:
      operationId: listClubs
      tags: [clubs]
      summary: List clubs at the user's university
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Clubs list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Club"
                required: [success, data]
  /clubs/{clubId}/join:
    post:
      operationId: joinClub
      tags: [clubs]
      summary: Request to join a club
      security:
        - bearerAuth: []
      parameters:
        - name: clubId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Join requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── OPPORTUNITIES ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ─── AUTH SCHEMAS ────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        language:
          type: string
          enum: [ar, fr, en]
          default: ar
      required: [fullName, email, password]
    LoginInput:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
      required: [accessToken, refreshToken, expiresIn]
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            tokens:
              $ref: "#/components/schemas/AuthTokens"
          required: [user, tokens]
      required: [success, data]
    TokenPairResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/AuthTokens"
      required: [success, data]
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/User"
      required: [success, data]

    # ─── USER SCHEMAS ─────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [student, professor, club_manager, department_admin, faculty_admin, university_admin, moderator, agent, finance_admin, super_admin]
        emailVerified:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        profile:
          $ref: "#/components/schemas/Profile"
          nullable: true
      required: [id, email, fullName, role, emailVerified, isActive, createdAt]
    Profile:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        universityId:
          type: string
          nullable: true
        facultyId:
          type: string
          nullable: true
        departmentId:
          type: string
          nullable: true
        levelId:
          type: string
          nullable: true
        groupId:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        privacy:
          type: string
          enum: [private, same_university, public]
        language:
          type: string
          enum: [ar, fr, en]
        university:
          $ref: "#/components/schemas/University"
          nullable: true
        faculty:
          $ref: "#/components/schemas/Faculty"
          nullable: true
        department:
          $ref: "#/components/schemas/Department"
          nullable: true
        level:
          $ref: "#/components/schemas/Level"
          nullable: true
        group:
          $ref: "#/components/schemas/Group"
          nullable: true
        onboardingComplete:
          type: boolean
      required: [id, userId, privacy, language, onboardingComplete]
...
    University:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        nameFr:
          type: string
          nullable: true
        city:
          type: string
        logoUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [community_created, verified, official_partner]
        facultyCount:
          type: integer
          nullable: true
      required: [id, name, city, status]
    Faculty:
      type: object
      properties:
        id:
          type: string
        universityId:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        nameFr:
          type: string
          nullable: true
      required: [id, universityId, name]
    Department:
      type: object
      properties:
        id:
          type: string
        facultyId:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        nameFr:
          type: string
          nullable: true
      required: [id, facultyId, name]
    Level:
      type: object
      properties:
        id:
          type: string
        departmentId:
          type: string
        name:
          type: string
        yearNumber:
          type: integer
      required: [id, departmentId, name, yearNumber]
    Group:
      type: object
      properties:
        id:
          type: string
        levelId:
          type: string
        name:
          type: string
      required: [id, levelId, name]

    # ─── COURSE SCHEMAS ───────────────────────────────────────────────────────
...
    Post:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        authorAvatarUrl:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [public, same_university, same_department, same_course, same_group]
        reactionCount:
          type: integer
        commentCount:
          type: integer
        userReaction:
          type: string
          nullable: true
        isPinned:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, content, authorId, authorName, visibility, reactionCount, commentCount, isPinned, createdAt]
    CreatePostInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        visibility:
          type: string
          enum: [public, same_university, same_department, same_course, same_group]
          default: same_university
      required: [content]

    # ─── EVENT SCHEMAS ────────────────────────────────────────────────────────
...
    Club:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        memberCount:
          type: integer
        isMember:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, memberCount, isMember, createdAt]

    # ─── OPPORTUNITY SCHEMAS ──────────────────────────────────────────────────
...